+++ codemods/debarrel/tests/package-name-alias-namespace-barrel/expected/src/icons.stories.tsx
@@ -1,0 +1,3 @@
+import * as Icons from "sentry/icons";
+
+console.log(Icons.IconAdd());