+++ codemods/debarrel/tests/package-name-alias-namespace-barrel/expected/src/stories/index.tsx
@@ -1,0 +1,2 @@
+export { story } from "./storybook";
+export { Demo } from "./demo";