+++ codemods/debarrel/tests/sentry-icons-namespace-barrel/input/static/app/icons/icons.stories.tsx
@@ -1,0 +1,4 @@
+import * as Icons from 'sentry/icons';
+
+// @ts-expect-error demo export
+export default Icons.IconAdd;