+++ codemods/debarrel/tests/sentry-static-app-namespace-barrels/expected/static/app/icons/icons.stories.tsx
@@ -1,0 +1,3 @@
+import * as Icons from 'sentry/icons';
+
+export default Icons.IconAdd;