+++ codemods/debarrel/tests/sentry-static-app-namespace-barrels/input/static/app/stories/index.tsx
@@ -1,0 +1,3 @@
+export {story} from './storybook';
+export {Demo} from './demo';
+export {StoryHeader} from './view/storyHeader';